+++ app/src/main/kotlin/org/michaelbel/movies/MainActivity.kt
@@ -13,13 +13,8 @@ import androidx.core.view.WindowCompat
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
-import androidx.work.OneTimeWorkRequestBuilder
-import androidx.work.WorkManager
-import androidx.work.workDataOf
 import dagger.hilt.android.AndroidEntryPoint
 import org.michaelbel.movies.common.theme.AppTheme
-import org.michaelbel.movies.domain.workers.AccountUpdateWorker
-import org.michaelbel.movies.domain.workers.MoviesDatabaseWorker
 import org.michaelbel.movies.ui.shortcuts.installShortcuts
 import org.michaelbel.movies.ui.theme.MoviesTheme
 
@@ -61,25 +56,5 @@ internal class MainActivity: AppCompatActivity() {
                 }
             }
         }
-
-        prepopulateDatabase()
-        updateAccountDetails()
-    }
-
-    private fun prepopulateDatabase() {
-        val request = OneTimeWorkRequestBuilder<MoviesDatabaseWorker>()
-            .setInputData(workDataOf(MoviesDatabaseWorker.KEY_FILENAME to MOVIES_DATA_FILENAME))
-            .build()
-        WorkManager.getInstance(this).enqueue(request)
-    }
-
-    private fun updateAccountDetails() {
-        val request = OneTimeWorkRequestBuilder<AccountUpdateWorker>()
-            .build()
-        WorkManager.getInstance(this).enqueue(request)
-    }
-
-    private companion object {
-        private const val MOVIES_DATA_FILENAME = "movies.json"
     }
 }
